Mobile workspace enhancement
