+++ project/FrontEnd/collaborative_science_platform/lib/screens/workspace_page/create_workspace_page/mobile_create_workspace_page.dart
