+++ project/FrontEnd/collaborative_science_platform/lib/screens/workspace_page/mobile_workspace_page/widget/mobile_workspace_content.dart
@@ -1,0 +1,265 @@
+import 'package:collaborative_science_platform/screens/workspace_page/mobile_workspace_page/widget/reference_card.dart';
+import 'package:collaborative_science_platform/screens/workspace_page/mobile_workspace_page/widget/subsection_title.dart';
+import 'package:flutter/material.dart';
+import 'package:provider/provider.dart';
+
+
+import '../../../../models/node.dart';
+import '../../../../models/user.dart';
+import '../../../../models/workspaces_page/entry.dart';
+import '../../../../models/workspaces_page/workspace.dart';
+import '../../../../providers/auth.dart';
+import '../../../../utils/lorem_ipsum.dart';
+import '../../../../utils/responsive/responsive.dart';
+import 'contributor_card.dart';
+import 'entry_card.dart';
+
+class MobileWorkspaceContent extends StatefulWidget {
+  final int workspaceId;
+  final bool pending;
+  const MobileWorkspaceContent({
+    super.key,
+    required this.workspaceId,
+    required this.pending,
+  });
+
+  @override
+  State<MobileWorkspaceContent> createState() => _MobileWorkspaceContentState();
+}
+
+class _MobileWorkspaceContentState extends State<MobileWorkspaceContent> {
+  bool isLoading = false;
+  bool error = false;
+  String errorMessage = "";
+
+  Workspace workspaceData = Workspace(
+    workspaceId: 0,
+    workspaceTitle: "workspaceTitle",
+    entries: <Entry>[],
+    status: WorkspaceStatus.workable,
+    numApprovals: 0,
+    contributors: <User>[],
+    pendingContributors: <User>[],
+    references: <Node>[],
+  );
+
+  @override
+  void didChangeDependencies() {
+    getWorkspaceData();
+    super.didChangeDependencies();
+  }
+
+  void getWorkspaceData() {
+    setState(() {
+      isLoading = true;
+    });
+    workspaceData = Workspace(
+      workspaceId: 0,
+      workspaceTitle: "workspaceTitle",
+      entries: <Entry>[
+        Entry(
+          content: getLongLoremIpsum(),
+          entryDate: DateTime.now(),
+          entryId: 1,
+          entryNumber: 1,
+          index: 1,
+          isEditable: false,
+          isFinalEntry: false,
+          isProofEntry: false,
+          isTheoremEntry: true,
+        ),
+        Entry(
+          content: getLongLoremIpsum(2),
+          entryDate: DateTime.now(),
+          entryId: 2,
+          entryNumber: 2,
+          index: 2,
+          isEditable: false,
+          isFinalEntry: false,
+          isProofEntry: true,
+          isTheoremEntry: false,
+        ),
+        Entry(
+          content: getLongLoremIpsum(3),
+          entryDate: DateTime.now(),
+          entryId: 2,
+          entryNumber: 2,
+          index: 2,
+          isEditable: false,
+          isFinalEntry: true,
+          isProofEntry: true,
+          isTheoremEntry: false,
+        ),
+      ],
+      status: WorkspaceStatus.workable,
+      numApprovals: 0,
+      contributors: <User>[
+        // Automatically add the user to the list of contributors
+        if (!widget.pending) Provider.of<Auth>(context).user as User,
+        User(
+          email: "dummy1@mail.com",
+          firstName: "dummy 1",
+          lastName: "jackson",
+        ),
+        User(
+          email: "dummy2@mail.com",
+          firstName: "dummy 2",
+          lastName: "jackson",
+        ),
+      ],
+      pendingContributors: <User>[
+        User(
+          email: "dummy3@mail.com",
+          firstName: "dummy 3",
+          lastName: "jackson",
+        ),
+      ],
+      references: <Node>[
+        Node(
+          contributors: <User>[],
+          id: 1,
+          nodeTitle: "Awesome Node Title",
+          publishDate: DateTime.now(),
+        ),
+      ],
+    );
+    setState(() {
+      isLoading = false;
+    });
+  }
+
+  Widget addIcon(Function() onPressed) {
+    return Center(
+      child: IconButton(
+        iconSize: 40.0,
+        onPressed: onPressed,
+        icon: const Icon(Icons.add),
+      ),
+    );
+  }
+
+  Widget firstAddition(String message, Function() onPressed) {
+    return ListView(
+      shrinkWrap: true,
+      physics: const NeverScrollableScrollPhysics(),
+      children: [
+        Center(
+          child: Text(
+            message,
+            style: const TextStyle(
+              fontSize: 16.0,
+              fontWeight: FontWeight.bold,
+            ),
+          ),
+        ),
+        addIcon(() {/* Navigate to a page where new entries are created */}),
+      ],
+    );
+  }
+
+  Widget entryList() {
+    int length = workspaceData.entries.length;
+    return workspaceData.entries.isNotEmpty
+      ? Padding(
+        padding: const EdgeInsets.only(bottom: 12.0),
+        child: ListView.builder(
+          padding: const EdgeInsets.all(0.0),
+          shrinkWrap: true,
+          physics: const NeverScrollableScrollPhysics(),
+          itemCount: length + 1,
+          itemBuilder: (context, index) => (index < length)
+            ? EntryCard(
+              entry: workspaceData.entries[index],
+              onDelete: () {
+                setState(() {
+                  workspaceData.entries.removeAt(index);
+                });
+              },
+              pending: widget.pending,
+            ) : addIcon(() {/* Navigate to a page where new entries are created */}),
+          ),
+        ) : firstAddition(
+          "Add Your First Entry!",
+          () {/* Navigate to a page where new entries are created */},
+    );
+  }
+
+  Widget contributorList() {
+    int length = workspaceData.contributors.length;
+    return workspaceData.contributors.isNotEmpty
+        ? Padding(
+          padding: const EdgeInsets.only(bottom: 12.0),
+          child: ListView.builder(
+            padding: const EdgeInsets.all(0.0),
+            shrinkWrap: true,
+            physics: const NeverScrollableScrollPhysics(),
+            itemCount: length + 1,
+            itemBuilder: (context, index) => (index < length)
+              ? ContributorCard(contributor: workspaceData.contributors[index])
+              : addIcon(() => {/* Navigate to a page where new contributors are added */}),
+            ),
+          ) : firstAddition(
+          "Add The First Contributor!",
+          () {/* Navigate to a page where new contributors are added */},
+    );
+  }
+
+  Widget referenceList() {
+    int length = workspaceData.references.length;
+    return (workspaceData.references.isNotEmpty)
+        ? Padding(
+          padding: const EdgeInsets.only(bottom: 12.0),
+          child: ListView.builder(
+            padding: const EdgeInsets.all(0.0),
+            shrinkWrap: true,
+            physics: const NeverScrollableScrollPhysics(),
+            itemCount: length + 1,
+            itemBuilder: (context, index) => (index < length)
+              ? ReferenceCard(reference: workspaceData.references[index])
+              : addIcon(() => {/* Navigate to a page where new references are added */}),
+            ),
+          ) : firstAddition(
+          "Add Your First Reference!",
+          () {/* Navigate to a page where new references are added */},
+    );
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    print("Created ${widget.workspaceId}");
+    if (isLoading || error) {
+      return Center(
+          child: isLoading ? const CircularProgressIndicator()
+              : error ? SelectableText(errorMessage)
+              : const SelectableText("Something went wrong!")
+      );
+    } else {
+      return SizedBox(
+        width: Responsive.getGenericPageWidth(context),
+        child: ListView(
+          shrinkWrap: true,
+          physics: const NeverScrollableScrollPhysics(),
+          padding: const EdgeInsets.all(0.0),
+          // It needs to be nested scrollable in the future
+          children: <Widget>[
+            const SizedBox(height: 10.0),
+            const SubSectionTitle(title: "Entries"),
+            entryList(),
+            const Padding(
+              padding: EdgeInsets.symmetric(horizontal: 12.0),
+              child: Divider(),
+            ),
+            const SubSectionTitle(title: "Contributors"),
+            contributorList(),
+            const Padding(
+              padding: EdgeInsets.symmetric(horizontal: 12.0),
+              child: Divider(),
+            ),
+            const SubSectionTitle(title: "References"),
+            referenceList(),
+          ],
+        ),
+      );
+    }
+  }
+}
