+++ project/FrontEnd/collaborative_science_platform/lib/screens/workspace_page/mobile_workspace_page/widget/reference_card.dart
