+++ project/FrontEnd/collaborative_science_platform/lib/screens/workspace_page/web_workspace_page/widgets/entry_form.dart
