+++ project/FrontEnd/collaborative_science_platform/lib/screens/workspace_page/web_workspace_page/widgets/references_list_view.dart
@@ -1,6 +1,6 @@
 import 'package:collaborative_science_platform/models/node.dart';
 import 'package:collaborative_science_platform/screens/node_details_page/node_details_page.dart';
-import 'package:collaborative_science_platform/screens/workspaces_page/web_workspace_page/widgets/add_reference_form.dart';
+import 'package:collaborative_science_platform/screens/workspace_page/web_workspace_page/widgets/add_reference_form.dart';
 import 'package:collaborative_science_platform/utils/text_styles.dart';
 import 'package:collaborative_science_platform/widgets/app_button.dart';
 import 'package:collaborative_science_platform/widgets/card_container.dart';
