+++ project/FrontEnd/collaborative_science_platform/lib/screens/workspace_page/web_workspace_page/widgets/send_collaboration_request_form.dart
