+++ project/FrontEnd/collaborative_science_platform/lib/screens/workspace_page/web_workspace_page/widgets/workspaces_side_bar.dart
@@ -1,7 +1,7 @@
 import 'package:collaborative_science_platform/models/workspaces_page/workspaces.dart';
 import 'package:collaborative_science_platform/models/workspaces_page/workspaces_object.dart';
 import 'package:collaborative_science_platform/screens/page_with_appbar/widgets/app_bar_button.dart';
-import 'package:collaborative_science_platform/screens/workspaces_page/web_workspace_page/widgets/create_workspace_form.dart';
+import 'package:collaborative_science_platform/screens/workspace_page/web_workspace_page/widgets/create_workspace_form.dart';
 import 'package:collaborative_science_platform/utils/colors.dart';
 import 'package:collaborative_science_platform/utils/text_styles.dart';
 import 'package:collaborative_science_platform/widgets/app_button.dart';
