+++ project/FrontEnd/collaborative_science_platform/lib/screens/workspace_page/workspaces_page.dart
@@ -1,8 +1,8 @@
-import 'package:collaborative_science_platform/screens/workspaces_page/mobile_workspaces_page.dart';
-import 'package:collaborative_science_platform/screens/workspaces_page/web_workspace_page/web_workspace_page.dart';
+import 'package:collaborative_science_platform/screens/workspace_page/web_workspace_page/web_workspace_page.dart';
 import 'package:flutter/material.dart';
 
 import '../../utils/responsive/responsive.dart';
+import 'mobile_workspace_page/mobile_workspace_page.dart';
 
 class WorkspacesPage extends StatelessWidget {
   static const routeName = '/workspaces';
@@ -11,7 +11,7 @@ class WorkspacesPage extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return const Responsive(
-      mobile: MobileWorkspacesPage(),
+      mobile: MobileWorkspacePage(),
       desktop: WebWorkspacePage(),
     );
   }
