--- project/FrontEnd/collaborative_science_platform/lib/screens/workspaces_page/web_workspaces_page.dart
@@ -1,15 +1,0 @@
-import 'package:flutter/material.dart';
-
-class WebWorkspacesPage extends StatefulWidget {
-  const WebWorkspacesPage({super.key});
-
-  @override
-  State<WebWorkspacesPage> createState() => _WebWorkspacesPageState();
-}
-
-class _WebWorkspacesPageState extends State<WebWorkspacesPage> {
-  @override
-  Widget build(BuildContext context) {
-    return const Placeholder();
-  }
-}
